requirements.txt
anaconda-anon-usage @ file:///C:/b/abs_e8r_zga7xy/croot/anaconda-anon-usage_1732732454901/work
annotated-types @ file:///C:/b/abs_0dmaoyhhj3/croot/annotated-types_1709542968311/work
archspec @ file:///croot/archspec_1709217642129/work
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
binaryornot==0.4.4
blinker==1.9.0
boltons @ file:///C:/b/abs_45_52ughkz/croot/boltons_1737061711836/work
branca==0.8.1
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==5.5.2
canvas==0.22.1
certifi @ file:///C:/b/abs_8a944p1_gn/croot/certifi_1738623753421/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
comm==0.2.2
conda @ file:///C:/b/abs_12ybmrg1d4/croot/conda_1738168389704/work
conda-anaconda-telemetry @ file:///C:/b/abs_4c9llcc5ob/croot/conda-anaconda-telemetry_1736524617431/work
conda-anaconda-tos @ file:///C:/b/abs_ceeuq0lee_/croot/conda-anaconda-tos_1739299022910/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1737733694612/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_bdz9vbvbh2/croot/conda-package-streaming_1731366449946/work
configparser==7.1.0
contourpy==1.3.1
cookiecutter==2.6.0
cron-converter==1.2.1
cryptography @ file:///C:/b/abs_e2lzchf4i6/croot/cryptography_1732130411942/work
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.9
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
Django==5.1.6
django-stubs==5.1.3
django-stubs-ext==5.1.3
django-timezone-utils==0.15.0
env-tools==2.4.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
folium==0.19.5
fonttools==4.56.0
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googlemaps==4.10.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httplib2==0.22.0
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.8
libmambapy @ file:///C:/b/abs_627vsv8bhu/croot/mamba-split_1734469608328/work/libmambapy
markdown-it-py @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markdown-it-py_1699473886965/work
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mdurl_1699473506455/work
menuinst @ file:///C:/b/abs_fblttj5gp1/croot/menuinst_1738943438301/work
more-itertools==10.6.0
mplcursors==0.6
mypy==1.15.0
mypy-extensions==1.0.0
nbformat==5.10.4
nbmerge==0.0.4
nest-asyncio==1.6.0
networkx==3.4.2
npm==0.1.1
numpy==2.2.4
optional-django==0.1.0
osmnx==2.0.3
outcome==1.3.0.post0
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
pipenv==2025.0.2
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
playwright==1.52.0
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg==3.2.5
psycopg-binary==3.2.5
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///C:/b/abs_27dx58x550/croot/pydantic_1734736090499/work
pydantic_core @ file:///C:/b/abs_bdosz7qwys/croot/pydantic-core_1734726071532/work
pyee==13.0.0
Pygments @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pygments_1699474141968/work
PyJWT==2.10.1
PyMySQL==1.1.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.1
qrcode==8.0
RapidFuzz==3.12.1
redis==5.2.1
referencing==0.36.2
reportlab==4.3.1
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
RestrictedPython==8.0
rich @ file:///C:/b/abs_8b92fzmygg/croot/rich_1732638991774/work
rpds-py==0.22.3
rsa==4.9.1
ruamel.yaml @ file:///C:/b/abs_0cunwx_ww6/croot/ruamel.yaml_1727980181547/work
ruamel.yaml.clib @ file:///C:/b/abs_5fk8zi6n09/croot/ruamel.yaml.clib_1727769837359/work
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.29.0
setuptools==75.8.0
shapely==2.1.0
shellingham==1.5.4
shutils==0.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
statsd==4.0.1
text-unidecode==1.3
threadpoolctl==3.6.0
tini==4.0.0
tornado==6.4.2
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
truststore @ file:///C:/b/abs_494cm143zh/croot/truststore_1736550137835/work
typer==0.15.1
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20241230
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.1
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
virtualenv==20.31.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wsproto==1.2.0
xyzservices==2025.1.0
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
